gateway_service/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
redis==6.2.0
sqlparse==0.5.3
